Remove mocks, reorganize types
